remove control model warning

+++ src/control/control.cpp
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-const int MAXN = 1e6 + 10;
+const int MAXN = (int)1e6 + 10;
 
 char* result[MAXN];
 char* wordList[MAXN_WORD];
